# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat/message": {
      "post": {
        "description": "发送一条文本消息并直接建立首轮 SSE 流；若未传 conversationId，则自动创建新会话",
        "operationId": "sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "聊天流式响应",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发送文本消息并开始流式回复",
        "tags": [
          "聊天"
        ]
      }
    },
    "/api/chat/voice-messages": {
      "post": {
        "description": "上传音频文件，Whisper 转文字后创建可恢复的流式任务；若未传 conversationId，则自动创建新会话",
        "operationId": "voiceCompletions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/VoiceCompletionsFormDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "语音聊天任务创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTaskResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "发送语音消息",
        "tags": [
          "聊天"
        ]
      }
    },
    "/api/chat/image-generations": {
      "post": {
        "description": "使用 DALL-E 3 生成图片",
        "operationId": "ChatController_imageGenerations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "图片生成结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerationResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "AI 图片生成",
        "tags": [
          "聊天"
        ]
      }
    },
...
  },
  "info": {
    "title": "办伴 Banban API",
    "description": "办伴 Banban 后端接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReasoningSelectionDto": {
        "type": "object",
        "properties": {
          "activation": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled",
              "auto"
            ]
          },
          "effort": {
            "type": "string",
            "enum": [
              "minimal",
              "low",
              "medium",
              "high",
              "xhigh",
              "max"
            ]
          },
          "budgetTokens": {
            "oneOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "string",
                "enum": [
                  "auto"
                ]
              }
            ]
          }
        }
      },
...
      "ChatCompletionsDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "description": "会话 ID；首轮消息可不传，后端会自动创建新会话",
            "example": "conv_abc123"
          },
          "content": {
            "type": "string",
            "description": "用户消息内容；与 imageAssetId 至少提供一项",
            "example": "请分析这张图片"
          },
          "imageAssetId": {
            "type": "string",
            "description": "本人已登记且 usage=chat-image 的图片资产 ID",
            "example": "cmf_asset_123"
          },
          "agentId": {
            "type": "string",
            "description": "指定使用的智能体 id；不传则用默认智能体",
            "example": "agent_abc123"
          },
          "selectedModelPresetId": {
            "type": "string",
            "description": "本条消息选择的模型预设业务 ID；必须属于回答智能体允许集合，仅替换 Flow 中的 agent-default",
            "example": "deepseek-official:deepseek-chat"
          },
          "reasoning": {
            "description": "仅 direct Agent 可用的本轮思考设置",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReasoningSelectionDto"
              }
            ]
          }
        }
      },
      "VoiceCompletionsFormDataDto": {
        "type": "object",
        "properties": {
          "audio": {
            "type": "string",
            "description": "音频文件",
            "format": "binary"
          },
          "conversationId": {
            "type": "string",
            "description": "会话 ID；首轮语音消息可不传，后端会自动创建新会话",
            "example": "cmf_conv_123"
          },
          "selectedModelPresetId": {
            "type": "string",
            "description": "本条消息选择的模型预设业务 ID；必须属于回答智能体允许集合，仅替换 Flow 中的 agent-default",
            "example": "deepseek-official:deepseek-chat"
          },
          "agentId": {
            "type": "string",
            "description": "指定使用的智能体 id；不传则使用会话或系统默认智能体",
            "example": "agent_abc123"
          },
          "reasoning": {
            "type": "string",
            "description": "JSON 字符串形式的本轮思考设置，仅 direct Agent 可用",
            "example": "{\"activation\":\"enabled\",\"effort\":\"high\"}"
          }
        },
        "required": [
          "audio"
        ]
      },
      "ChatTaskResultDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "description": "流式任务 ID",
            "example": "cmf_task_123"
          },
          "streamId": {
            "type": "string",
            "description": "当前执行流片段 ID",
            "example": "cmf_stream_123"
          },
          "messageId": {
            "type": "string",
            "description": "本次 assistant 占位消息 ID",
            "example": "cmf_msg_123"
          },
          "conversationId": {
            "type": "string",
            "description": "会话 ID",
            "example": "cmf_conv_123"
          },
          "status": {
            "type": "string",
            "description": "任务初始状态",
            "enum": [
              "pending",
              "streaming",
              "paused",
              "waiting_human",
              "completed",
              "error",
              "expired",
              "canceled"
            ],
            "example": "pending"
          }
        },
        "required": [
          "taskId",
          "streamId",
          "messageId",
          "conversationId",
          "status"
        ]
      },
      "ImageGenerationDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "description": "会话 ID",
            "example": "conv_abc123"
          },
          "prompt": {
            "type": "string",
            "description": "图片生成提示词",
            "example": "一只可爱的小熊在森林里散步",
            "maxLength": 4000
          }
        },
        "required": [
          "conversationId",
          "prompt"
        ]
      },
      "ImageGenerationResultDto": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string",
            "description": "图片消息 ID",
            "example": "cmf_msg_123"
          },
          "imageUrl": {
            "type": "string",
            "description": "生成图片地址",
            "example": "https://example.com/generated-image.png"
          },
          "revisedPrompt": {
            "type": "string",
            "description": "模型修订后的提示词",
            "example": "一只可爱的小熊在森林里散步"
          }
        },
        "required": [
          "messageId",
          "imageUrl",
          "revisedPrompt"
        ]
      }
    }
  }
}